Restructure components. Add grid styling.
