+++ client/src/components/CalendarMonth/CalendarMonth.module.css
@@ -1,0 +1,7 @@
+.calendar {
+  display: grid;
+  grid-template-columns: repeat(7, 50px);
+  grid-template-rows: 30px;
+  grid-auto-rows: 50px;
+  place-items: center;
+}
